requirements.txt
youtube-transcript-api>=1.1.0
openpyxl>=3.1.5
google-genai>=1.21.1
requests>=2.32.4
python-dotenv>=1.1.1